# webChat.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "webChat",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "friend"
    },
    {
      "name": "user"
    },
    {
      "name": "take"
    },
    {
      "name": "group"
    },
    {
      "name": "message"
    },
    {
      "name": "group message"
    }
  ],
  "paths": {
    "/friend/user/{id}": {
      "get": {
        "summary": "获取用户的朋友",
        "x-apifox-folder": "friend",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "friend"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137765461-run",
        "security": []
      }
    },
    "/user/login": {
      "post": {
        "summary": "登录",
        "x-apifox-folder": "user",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "x-apifox-orders": [
                  "name",
                  "password"
                ],
                "required": [
                  "name",
                  "password"
                ],
                "x-apifox-ignore-properties": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137766903-run",
        "security": []
      }
    },
    "/user/signup": {
      "post": {
        "summary": "注册",
        "x-apifox-folder": "user",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "password"
                ],
                "x-apifox-orders": [
                  "name",
                  "password"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": {
                "name": "string",
                "password": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137766914-run",
        "security": []
      }
    },
    "/friend": {
      "post": {
        "summary": "加好友",
        "x-apifox-folder": "friend",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "friend"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "friendName",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137768489-run",
        "security": []
      },
      "delete": {
        "summary": "删好友",
        "x-apifox-folder": "friend",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "friend"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "friendId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137772410-run",
        "security": []
      }
    },
    "/user/{id}": {
      "get": {
        "summary": "获取用户信息",
        "x-apifox-folder": "user",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137771079-run",
        "security": []
      }
    },
    "/take/user/{id}": {
      "get": {
        "summary": "获取用户的群",
        "x-apifox-folder": "take",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "take"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137771185-run",
        "security": []
      }
    },
    "/group/{groupId}": {
      "get": {
        "summary": "获取群信息",
        "x-apifox-folder": "group",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "group"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137771334-run",
        "security": []
      }
    },
    "/message": {
      "post": {
        "summary": "发送消息",
        "x-apifox-folder": "message",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "message"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "msgContent": {
                    "type": "string"
                  },
                  "msgFrom": {
                    "type": "integer"
                  },
                  "msgTo": {
                    "type": "integer"
                  }
                },
                "x-apifox-orders": [
                  "msgContent",
                  "msgFrom",
                  "msgTo"
                ],
                "required": [
                  "msgContent",
                  "msgFrom",
                  "msgTo"
                ],
                "x-apifox-ignore-properties": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137772039-run",
        "security": []
      },
      "delete": {
        "summary": "删除信息",
        "x-apifox-folder": "message",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "message"
        ],
        "parameters": [
          {
            "name": "msgId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137772140-run",
        "security": []
      }
    },
    "/groupMessage": {
      "post": {
        "summary": "发送群消息",
        "x-apifox-folder": "group message",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "group message"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "groupMsgContent": {
                    "type": "string"
                  },
                  "groupMsgFrom": {
                    "type": "integer"
                  },
                  "groupId": {
                    "type": "integer"
                  }
                },
                "x-apifox-orders": [
                  "groupMsgContent",
                  "groupMsgFrom",
                  "groupId"
                ],
                "required": [
                  "groupMsgContent",
                  "groupMsgFrom",
                  "groupId"
                ],
                "x-apifox-ignore-properties": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137772050-run",
        "security": []
      },
      "delete": {
        "summary": "删除群信息",
        "x-apifox-folder": "group message",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "group message"
        ],
        "parameters": [
          {
            "name": "groupMsgId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137772161-run",
        "security": []
      }
    },
    "/user": {
      "put": {
        "summary": "修改用户信息",
        "x-apifox-folder": "user",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1
                  },
                  "avatar": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "signature": {
                    "type": "string"
                  }
                },
                "x-apifox-orders": [
                  "id",
                  "name",
                  "gender",
                  "avatar",
                  "password",
                  "signature"
                ],
                "required": [
                  "id"
                ],
                "x-apifox-ignore-properties": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137772097-run",
        "security": []
      }
    },
    "/message/user": {
      "get": {
        "summary": "获取两个用户间消息",
        "x-apifox-folder": "message",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "message"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "friendId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137772118-run",
        "security": []
      }
    },
    "/groupMessage/group": {
      "get": {
        "summary": "获取群消息",
        "x-apifox-folder": "group message",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "group message"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3838394/apis/api-137772160-run",
        "security": []
      }
    }
  },
  "components": {
    "schemas": {},
    "securitySchemes": {}
  },
  "servers": []
}